# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Crown Gift Management API",
    "version": "1.0.0",
    "description": "This is a backend API for managing crown stock issue management system",
    "contact": {
      "name": "Mehedi Hassan",
      "email": "mehedi@example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "StockItem": {
        "type": "object",
        "required": [
          "name",
          "description",
          "quantity"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "description": "Stock item ID"
          },
          "name": {
            "type": "string",
            "description": "Stock item name"
          },
          "description": {
            "type": "string",
            "description": "Stock item description"
          },
          "quantity": {
            "type": "number",
            "minimum": 0,
            "description": "Available quantity"
          },
          "category": {
            "type": "string",
            "description": "Item category"
          },
          "unit": {
            "type": "string",
            "description": "Unit of measurement"
          }
        }
      },
      "StockIssue": {
        "type": "object",
        "required": [
          "stockItemId",
          "receiverId",
          "quantity"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "description": "Stock issue ID"
          },
          "stockItemId": {
            "type": "string",
            "description": "Reference to stock item"
          },
          "receiverId": {
            "type": "string",
            "description": "Reference to receiver user"
          },
          "quantity": {
            "type": "number",
            "minimum": 1,
            "description": "Quantity to issue"
          },
          "issuedDate": {
            "type": "string",
            "format": "date-time",
            "description": "Issue date"
          },
          "notes": {
            "type": "string",
            "description": "Additional notes"
          }
        }
      },
      "StockModification": {
        "type": "object",
        "required": [
          "stockItemId",
          "quantityChange",
          "reason"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "description": "Stock modification ID"
          },
          "stockItemId": {
            "type": "string",
            "description": "Reference to stock item"
          },
          "quantityChange": {
            "type": "number",
            "description": "Quantity change (positive for addition, negative for reduction)"
          },
          "reason": {
            "type": "string",
            "description": "Reason for modification"
          },
          "modifiedDate": {
            "type": "string",
            "format": "date-time",
            "description": "Modification date"
          },
          "modifiedBy": {
            "type": "string",
            "description": "User who made the modification"
          }
        }
      },
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "error"
          },
          "message": {
            "type": "string"
          },
          "error": {
            "type": "object"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/stock-issues/stockItem/{stockItemId}": {
      "get": {
        "summary": "Get stock issues by stock item ID",
        "tags": [
          "Stock Issues"
        ],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "stockItemId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Stock item ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Stock issues retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/StockIssue"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/stock-items": {
      "get": {
        "summary": "Get all stock items with pagination",
        "tags": [
          "Stock Items"
        ],
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            },
            "description": "Number of items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Stock items retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/StockItem"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create stock item(s)",
        "tags": [
          "Stock Items"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/StockItem"
                  },
                  {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/StockItem"
                    }
                  }
                ]
              },
              "examples": {
                "single": {
                  "summary": "Create single stock item",
                  "value": {
                    "name": "Laptop",
                    "description": "Dell Laptop 15 inch",
                    "quantity": 10,
                    "category": "Electronics",
                    "unit": "piece"
                  }
                },
                "multiple": {
                  "summary": "Create multiple stock items",
                  "value": [
                    {
                      "name": "Laptop",
                      "description": "Dell Laptop 15 inch",
                      "quantity": 10,
                      "category": "Electronics",
                      "unit": "piece"
                    },
                    {
                      "name": "Mouse",
                      "description": "Wireless Mouse",
                      "quantity": 25,
                      "category": "Electronics",
                      "unit": "piece"
                    }
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Stock item(s) created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "oneOf": [
                            {
                              "$ref": "#/components/schemas/StockItem"
                            },
                            {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/StockItem"
                              }
                            }
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin, Issuer, or Approver role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/stock-items/{id}": {
      "get": {
        "summary": "Get stock item by ID",
        "tags": [
          "Stock Items"
        ],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Stock item ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Stock item retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/StockItem"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Stock item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update stock item",
        "tags": [
          "Stock Items"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Stock item ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "quantity": {
                    "type": "number",
                    "minimum": 0
                  },
                  "category": {
                    "type": "string"
                  },
                  "unit": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stock item updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/StockItem"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin, Issuer, or Approver role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Stock item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete stock item",
        "tags": [
          "Stock Items"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Stock item ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Stock item deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin, Issuer, or Approver role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Stock item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/stock-modifications/stock-item/{stockItemId}": {
      "get": {
        "summary": "Get stock modifications by stock item ID",
        "tags": [
          "Stock Modifications"
        ],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "stockItemId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Stock item ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Stock modifications retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/StockModification"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}